Ran inventory-management-system code
Ran code in the inventory-management-system folder which was copied from the source repo. Was able to successfully open and use the app. I had to change the last cm to `python -m flask run`  to get it running though. I think we could use the example's inventory database code for our pantry and sore inventory information. The html and css templates will also work well as a base for our design.