Create recipe_picker.py
The code we were basing our project on was not woking as we had hoped. We have developed this code as a simplified version. It does not take user input to create the recipie lists. The user can choose from recipies already set in the code. This code is original and was not taken from another  users repo

Co-Authored-By: Ally <69776903+ahamil53@users.noreply.github.com>